requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.2.0
attrs==23.2.0
certifi==2023.11.17
charset-normalizer==3.3.2
decorator==5.1.1
distro==1.9.0
frozenlist==1.4.1
h11==0.14.0
httpcore==1.0.2
httpx==0.26.0
idna==3.6
multidict==6.0.4
numpy==1.26.3
openai==1.9.0
py==1.11.0
pydantic==2.5.3
pydantic_core==2.14.6
python-dotenv==1.0.1
python-graphql-client==0.4.3
requests==2.31.0
retry==0.9.2
sniffio==1.3.0
tqdm==4.66.1
typing_extensions==4.9.0
urllib3==2.1.0
websockets==12.0
yarl==1.9.4
